pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10.0,<3.13"
google-generativeai = "^0.3.1"
xds-protos = "^1.60.0"
markdown = "^3.10"
pymupdf = "^1.26.6"
python-docx = "^1.2.0"
pdfplumber = "^0.11.8"
docutils = "^0.22.3"
pytest-asyncio = "^1.3.0"
hypothesis = "^6.148.7"
pytest = "^9.0.1"
asyncpg = "^0.31.0"
psycopg2-binary = "^2.9.11"
anthropic = "^0.75.0"
google-genai = "^1.53.0"
openai = "^2.9.0"
tenacity = "^9.1.2"
fastapi = "^0.124.0"
uvicorn = "^0.38.0"
python-multipart = "^0.0.20"
litellm = "^1.80.8"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"